Handle notifications about unknown payments
Per PayU protocol, We need to receive a notification even if we can't
recognize the particular payment. We save the reported payment data,
but this payment is not assigned to any particular user. Thus we must
excpect to have payments without any user assigned and the
administration must check and handle them manually.