apps/dpnk: fix PayU order product unit price and total amount
Specify prices using the lowest currency unit. For instance, in the case
of PLN (Polish Zloty), the lowest currency unit is the penny, which
equals 1/100 PLN. Therefore, if the price is 1000 in the lowest currency
unit, it would be equivalent to 10 PLN. However, HUF (Hungarian Forint)
is an exception to this rule. For HUF, you should multiply the amount
by 100.

https://developers.payu.com/europe/docs/payment-flows/auth-and-order/